Move custom analysis options into sub-packages (#4603)
When the legacy analysis options were bulk-added to the repository to
allow transitioning incrementally to the Flutter analysis options, they
were added to top-level plugin groups. This makes it harder to migrate
incrementally, and also means that when we add a new sub-package (e.g, a
web or desktop implementation) it's automatically using the legacy
options.

This moves all the legacy option files into each individual package, and
updates the allow list for the tool accordingly, so that migration can
be done package by package.